Refactor the initial mapping logic into a utils file.
